# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/question-generation/tasks": {
      "get": {
        "operationId": "QuestionGenerationController_list",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGeneration"
        ]
      },
      "post": {
        "operationId": "QuestionGenerationController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGeneration"
        ]
      }
    },
    "/api/user/question-generation/tasks/{taskId}/start": {
      "post": {
        "operationId": "QuestionGenerationController_start",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGeneration"
        ]
      }
    },
    "/api/user/question-generation/tasks/{taskId}/snapshot": {
      "get": {
        "operationId": "QuestionGenerationController_snapshot",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGeneration"
        ]
      },
      "put": {
        "operationId": "QuestionGenerationController_saveSnapshot",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGeneration"
        ]
      }
    },
    "/api/user/question-generation/tasks/{taskId}/questions/batch": {
      "post": {
        "operationId": "QuestionGenerationController_persistDrafts",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGeneration"
        ]
      }
    },
    "/api/user/question-generation/tasks/{taskId}/questions/approve": {
      "patch": {
        "operationId": "QuestionGenerationController_approve",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGeneration"
        ]
      }
    },
    "/api/user/question-generation/tasks/{taskId}/questions/{questionId}": {
      "patch": {
        "operationId": "QuestionGenerationController_updateDraft",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "questionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGeneration"
        ]
      }
    },
    "/api/user/question-generation/tasks/{taskId}": {
      "delete": {
        "operationId": "QuestionGenerationController_remove",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGeneration"
        ]
      }
    },
    "/api/v1/question-generation/tasks": {
      "get": {
        "operationId": "QuestionGenerationV1Controller_list",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGenerationV1"
        ]
      },
      "post": {
        "operationId": "QuestionGenerationV1Controller_create",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGenerationV1"
        ]
      }
    },
    "/api/v1/question-generation/tasks/{id}/start": {
      "post": {
        "operationId": "QuestionGenerationV1Controller_start",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGenerationV1"
        ]
      }
    },
    "/api/v1/question-generation/tasks/{id}/snapshot": {
      "get": {
        "operationId": "QuestionGenerationV1Controller_snapshot",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGenerationV1"
        ]
      },
      "put": {
        "operationId": "QuestionGenerationV1Controller_saveSnapshot",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGenerationV1"
        ]
      }
    },
    "/api/v1/question-generation/tasks/{id}/questions/batch": {
      "post": {
        "operationId": "QuestionGenerationV1Controller_persist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGenerationV1"
        ]
      }
    },
    "/api/v1/question-generation/tasks/{id}/questions/approve": {
      "patch": {
        "operationId": "QuestionGenerationV1Controller_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGenerationV1"
        ]
      }
    },
    "/api/v1/question-generation/tasks/{id}/questions/{questionId}": {
      "patch": {
        "operationId": "QuestionGenerationV1Controller_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "questionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGenerationV1"
        ]
      }
    },
    "/api/v1/question-generation/tasks/{id}": {
      "delete": {
        "operationId": "QuestionGenerationV1Controller_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "QuestionGenerationV1"
        ]
      }
    },
...
    "/api/user/notifications/unread": {
      "get": {
        "operationId": "NotificationController_getUnread",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Notification"
        ]
      }
    },
...
    "/api/user/notifications/{id}/read": {
      "post": {
        "operationId": "NotificationController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Notification"
        ]
      }
    },
...
    "/api/v1/notifications/unread-count": {
      "get": {
        "operationId": "NotificationV1Controller_unreadCount",
        "parameters": [
          {
            "name": "X-Client-App",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "读取当前用户未读通知数",
        "tags": [
          "v1 notifications"
        ]
      }
    },
    "/api/v1/notifications/unread": {
      "get": {
        "operationId": "NotificationV1Controller_unread",
        "parameters": [
          {
            "name": "X-Client-App",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "v1 notifications"
        ]
      }
    },
    "/api/v1/notifications": {
      "get": {
        "operationId": "NotificationV1Controller_list",
        "parameters": [
          {
            "name": "X-Client-App",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "v1 notifications"
        ]
      }
    },
    "/api/v1/notifications/{id}/read": {
      "post": {
        "operationId": "NotificationV1Controller_read",
        "parameters": [
          {
            "name": "X-Client-App",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "v1 notifications"
        ]
      }
    },
    "/api/v1/notifications/read-all": {
      "post": {
        "operationId": "NotificationV1Controller_readAll",
        "parameters": [
          {
            "name": "X-Client-App",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "v1 notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "ZhiPath Platform API",
    "description": "智途 ZhiPath、CodeNova 与后续定制客户端共享平台 API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "client-app": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Client-App"
      }
    },
...
  }
}